MSBS-276
Change the blue “change calendar category products” to show the select product category so we see what is selected.

 Allow for an appointment to be added by selecting on the actual calendar slot.

Allow the calendar view to move to the previous day using the arrows.

The calendar should show full screen.

We should allow for double booking.

Appointment info should not be cut out.